+++ Engine/Source/Runtime/RenderCore/Public/PixelShaderUtils.h
@@ -65,7 +65,7 @@ struct RENDERCORE_API FPixelShaderUtils
 		uint32 StencilRef = 0)
 	{
 		check(PixelShader.IsValid());
-		RHICmdList.SetViewport(Viewport.Min.X, Viewport.Min.Y, 0.0f, Viewport.Max.X, Viewport.Max.Y, 1.0f);
+		RHICmdList.SetViewport((float)Viewport.Min.X, (float)Viewport.Min.Y, 0.0f, (float)Viewport.Max.X, (float)Viewport.Max.Y, 1.0f);
 		
 		FGraphicsPipelineStateInitializer GraphicsPSOInit;
 		InitFullscreenPipelineState(RHICmdList, GlobalShaderMap, PixelShader, /* out */ GraphicsPSOInit);
